+++ test/app_crypto_icons_test.dart
@@ -1,7 +1,1 @@
-import 'package:flutter/services.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:app_crypto_icons/app_crypto_icons.dart';
-
-void main() {
-  
-}
+void main() {}