+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.0.1
+## 0.0.2
+* Added new image source type, url
 
-* TODO: Describe initial release.
+## 0.0.1
+* Initialize Package